test(audit): consolidate the two auth-audit integration files into one
auditAuth resolves the actor from the real login flow (session, req.login,
pending-MFA, JWT reset, WorkOS SSO 302), so all cases stay integration. Fold
audit.auth.managed into audit.auth.private as one 'audit — auth flows' suite
(local + managed/SSO), delete the managed file, and assert accountId.
